(bug) Fix <Dome> children keys to fix new positions
Formerly, adding a photo via an upload would trigger a re-render
with the new photos fetched from the servers, as it should, but
the positions and angles of the photos all got messed up because
of this. So changing the way each <Image> gets assigned a key prop
based on the image URL instead of its index fixes this.